# openapi.yml
openapi: 3.0.0
info:
  title: Library Management Application
  description: Demo application for managing a library
  version: 2.0.0
paths:
...
  /books:
    get:
      tags:
        - Books
      summary: Return a list of books
      responses:
        '200':
          description: The list of books as array of objects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/bookResponse'
    post:
      tags:
        - Books
      summary: Create a book
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/bookRequestBody'
      responses:
        '201':
          description: The book object created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/bookResponse'
        '400':
          description: If validation fails, a Bad Request HTTP status will be returned instead
  /books/{bookId}:
    parameters:
      - $ref: '#/components/parameters/bookId'
    get:
      tags:
        - Books
      summary: Return a book by id
      responses:
        '200':
          description: An object of the book
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/bookResponse'
        '404':
          description: If the bookId doesn't exists, a Not Found HTTP status will be returned instead
    put:
      tags:
        - Books
      summary: Update a book
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/bookRequestBody'
      responses:
        '204':
          description: If the book has been successfully updated, a No Content HTTP status will be returned
        '400':
          description: If validation fails, a Bad Request HTTP status will be returned instead
        '404':
          description: If the bookId doesn't exists, a Not Found HTTP status will be returned instead
    delete:
      tags:
        - Books
      summary: Delete a book
      responses:
        '204':
          description: If the book has been successfully removed, a No Content HTTP status will be returned
        '404':
          description: If the bookId doesn't exists, a Not Found HTTP status will be returned instead
components:
  parameters:
...
    bookId:
      name: bookId
      in: path
      required: true
      schema:
        type: integer
  schemas:
...
    bookResponse:
      type: object
      properties:
        bookId:
          type: integer
          readOnly: true
        title:
          type: string
        creationDate:
          type: string
          format: datetime
        updatedDate:
          type: string
          format: datetime
      example:
        bookId: 1
        title: 'The Present'
        creationDate: '2023-01-01 00:00:00'
        updatedDate: '2023-01-01 00:00:00'
...
    bookRequestBody:
      type: object
      properties:
        title:
          type: string
      example:
        title: 'The Present'
...